requirements.txt
--extra-index-url=https://download.pytorch.org/whl/cpu
torchaudio==2.4.1
transformers[onnx]==4.44.2
optimum[onnxruntime]==1.22.0
funasr_onnx==0.4.1
pytube==15.0.0
datasets
flask
librosa
modelscope==1.18.1
pysrt==1.1.2
onnxruntime==1.19.2
jieba==0.42.1
resampy==0.4.3
psutil
pytubefix
funasr
pytest